Adding doc tests to Data class
Testing selection of articles for the DbD site.
Adding doc tests to Article and SassBuilder and TextileAdapter
Adding background color to DbD featured component
Remove faulty  tag in article.
Add real items method to Data class. This has some implication for
callers of data.items, since that that now is a list of key-value
pairs. Needs to be adjusted in the callers.